Frontend refactor to move response handler, key input handling, and more into new utilities folder
Part of #124
